Adjust the way advanced properties works to allow additional advanced properties.  Also, add a zoom-in (+), zoom-out (-) capability to the global view.
git-svn-id: http://chianti.ucsd.edu/svn@24716 0ecc0d97-ab19-0410-9704-bfe1a75892f5

+++ csplugins/trunk/ucsf/scooter/clusterMaker/src/clusterMaker/algorithms/EdgeAttributeHandler.java
@@ -102,7 +102,7 @@ public class EdgeAttributeHandler
 
 		clusterProperties.add(new Tunable("attributeListGroup",
 		                                  "Source for array data",
-		                                  Tunable.GROUP, new Integer(6)));
+		                                  Tunable.GROUP, new Integer(5)));
 
 		// The attribute to use to get the weights
 		attributeArray = getAllAttributes();
